+++ src/index.ts
@@ -3,13 +3,10 @@ export { IncomingMessage as IncomingMessage } from 'http';
 export {
   AuthenticateOptions,
   AuthenticatePayload,
+  AuthenticationResponse,
+  ErrorBody,
   ErrorResponse,
 } from './common';
-
-export {
-  InvalidGrantTypeError,
-  UnsupportedGrantTypeError,
-} from './errors';
 
 export {
   default as Instance,
@@ -18,7 +15,4 @@ export {
 
 export { default as BaseClient } from './base_client';
 
-export {
-  AuthenticationResponse,
-  TokenWithExpiry,
-} from './authenticator';
+export { TokenWithExpiry } from './authenticator';